+++ src/p_mobj.c
@@ -1287,7 +1287,7 @@ mobj_t *P_SpawnMapThing(mapthing_t *mthing, bool spawnmonsters)
 
     // [BH] initialize certain mobj's animations to a random start frame
     // so groups of same mobjs are deliberately out of sync with each other
-    if (info->frames > 1)
+    if (info->frames > 1 && r_randomstartframes)
     {
         const int   frames = M_BigRandomInt(0, info->frames);
         state_t     *st = mobj->state;